requirements/development.txt
amqp==2.3.2
asn1crypto==0.24.0
astroid==2.0.4
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
coverage==4.5.1
cryptography==2.3.1
Django==2.1.1
django-redis==4.9.0
django-rest-knox==3.1.5
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.1.1
pyOpenSSL==18.0.0
pytz==2018.5
redis==2.10.6
six==1.11.0
vine==1.1.4
wrapt==1.10.11
dj-database-url==0.5.0
psycopg2-binary
